requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

#-i https://pypi.org/simple
asgiref             == 3.5.2
Django              == 4.1.3
django-filter       == 22.1
djangorestframework == 3.14.0
mock                == 4.0.3
Pillow              == 9.3.0
pip                 == 22.3.1
pytz                == 2022.6
setuptools          == 65.5.1
sqlparse            == 0.4.3
wheel               == 0.37.1


